Use `@http` instead of `@json` for `dryRun` parameter
Currently we are using the `@json` annotation to force the name of the
dry run query parameter to `dryRun`. But this is incorrect, and only
works because of a bug in the metamodel that is going to be fixed. This
patch replaces `@json` with `@http` which will work after the bug is
fixed.

Related: https://github.com/openshift-online/ocm-api-metamodel/pull/197
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>