OCM-209 | feat: add autoscaler API resources
After discussing about it, we decided that instead of handling all
cluster-autoscaler parameters via requests to the ``cluster`` resource,
we should allow the editing functionality only by the handling of the
``/api/clusters_mgmt/v1/clusters/{cluster_id}/autoscaler`` endpoint.

We're still interested in setting autoscaler parameters in POST requests
to the cluster object, as it's a great help to some consumers (most
importantly the OCM UI).

This means:
* We need to add a resource file in the meta-model, to make sure the SDK
  has a way to build requests against it. The URI enables ``POST, GET,
  DELETE, PUT`` for the singleton resource
  (``/api/clusters_mgmt/v1/clusters/{cluster_id}/autoscaler``. No
  special ID it allocated for the autoscaler resource).
* The autoscaler is no longer a struct but rather a class, as it now has
  an actual identity.
* The ``cluster`` class only contains a link to the ``autoscaler``
  resource. This was made to spare doing database joins, as well as
  making it more clear to API users that setting cluster-autoscaler
  values is *not* possible when doing PATCH requests to the cluster API
  resource.